refactor(imagemin): renamed `imagesmin` to `imagemin`
would be breaking change if feature was documented for consumer, but currently isn't so I'm regarding this a simple refactor.
