+++ lib/parse-imagemin.js
